+++ app/javascript/components/receipts/TransactionForm.jsx
