+++ app/javascript/components/requests/Request.jsx
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import React from "react";
 import StyledModal from "../helpers/StyledModal";
-import TransactionForm from "../transactions/TransactionForm";
+import TransactionForm from "../receipts/TransactionForm";
 import BuyerSnapshot from "../buyers/BuyerSnapshot";
 import ArtistSnapshot from "../artists/ArtistSnapshot";
 import WorkFixedPanel from "../works/WorkFixedPanel";
